@@ -1,6 +1,6 @@
 {
   "name": "odata-v4-server",
-  "version": "0.2.12",
+  "version": "0.2.13",
   "description": "OData V4 Server",
   "main": "build/lib/index.js",
   "typings": "build/lib/index",
@@ -79,61 +79,61 @@
   },
   "homepage": "https://github.com/jaystack/odata-v4-server#readme",
   "dependencies": {
-    "@types/body-parser": "0.0.33",
-    "@types/cors": "0.0.33",
-    "@types/express": "^4.11.0",
+    "@types/body-parser": "1.17.0",
+    "@types/cors": "2.8.4",
+    "@types/express": "^4.16.0",
     "@types/qs": "^6.5.1",
-    "body-parser": "^1.18.2",
+    "body-parser": "^1.18.3",
     "cors": "^2.8.4",
-    "deepmerge": "^1.5.2",
-    "express": "^4.16.2",
+    "deepmerge": "^2.1.1",
+    "express": "^4.16.3",
     "odata-v4-literal": "^0.1.1",
     "odata-v4-metadata": "^0.1.5",
     "odata-v4-parser": "git+https://github.com/CPGToolbox/odata-v4-parser",
     "odata-v4-service-document": "0.0.3",
     "odata-v4-service-metadata": "^0.1.6",
-    "qs": "^6.5.1",
+    "qs": "^6.5.2",
     "reflect-metadata": "^0.1.12",
-    "tslib": "^1.9.0"
+    "tslib": "^1.9.3"
   },
   "devDependencies": {
     "@types/benchmark": "^1.0.31",
-    "@types/deepmerge": "^1.3.3",
-    "@types/event-stream": "^3.3.33",
+    "@types/deepmerge": "^2.1.0",
+    "@types/event-stream": "^3.3.34",
     "@types/jsonstream": "^0.8.30",
-    "@types/lodash": "^4.14.98",
-    "@types/mocha": "^2.2.47",
-    "@types/mongodb": "^2.2.20",
-    "@types/node": "^7.0.52",
-    "@types/request": "^2.47.0",
-    "@types/request-promise": "^4.1.41",
+    "@types/lodash": "^4.14.112",
+    "@types/mocha": "^5.2.5",
+    "@types/mongodb": "^3.1.1",
+    "@types/node": "^10.5.2",
+    "@types/request": "^2.47.1",
+    "@types/request-promise": "^4.1.42",
     "@types/stream-buffers": "^3.0.2",
-    "JSONStream": "^1.3.2",
-    "benchmark": "^2.1.1",
-    "chai": "^3.5.0",
-    "copyfiles": "^1.2.0",
+    "JSONStream": "^1.3.3",
+    "benchmark": "^2.1.4",
+    "chai": "^4.1.2",
+    "copyfiles": "^2.0.0",
     "event-stream": "^3.3.4",
     "istanbul": "^1.1.0-alpha.1",
-    "mocha": "^3.5.3",
-    "mochawesome": "^1.5.3",
-    "mongodb": "^2.2.34",
-    "mssql": "^3.3.0",
-    "npm-watch": "^0.1.6",
-    "nyc": "^11.4.1",
+    "mocha": "^5.2.0",
+    "mochawesome": "^3.0.2",
+    "mongodb": "^3.1.1",
+    "mssql": "^4.1.0",
+    "npm-watch": "^0.3.0",
+    "nyc": "^12.0.2",
     "odata-v4-inmemory": "^0.1.9",
     "odata-v4-mongodb": "^0.1.12",
-    "remap-istanbul": "^0.9.6",
-    "request": "^2.85.0",
+    "remap-istanbul": "^0.11.1",
+    "request": "^2.87.0",
     "request-promise": "^4.2.2",
     "rimraf": "^2.6.2",
-    "source-map-support": "^0.4.18",
-    "stream-buffers": "^3.0.1",
-    "ts-node": "^4.1.0",
-    "tslint": "^5.9.1",
-    "typedoc": "^0.9.0",
-    "typedoc-default-themes": "^0.4.0",
-    "typedoc-plugin-external-module-name": "^1.0.10",
-    "typescript": "^2.7.1",
-    "xml-beautifier": "^0.3.0"
+    "source-map-support": "^0.5.6",
+    "stream-buffers": "^3.0.2",
+    "ts-node": "^7.0.0",
+    "tslint": "^5.11.0",
+    "typedoc": "^0.11.1",
+    "typedoc-default-themes": "^0.5.0",
+    "typedoc-plugin-external-module-name": "^1.1.1",
+    "typescript": "^2.9.2",
+    "xml-beautifier": "^0.4.0"
   }
 }
