@@ -1577,7 +1577,11 @@ export class ODataProcessor extends Transform {
             if (typeof queryAst == "string") {
                 queryAst = this.serverType.parser.query(queryAst, { metadata: this.resourcePath.ast.metadata || this.serverType.$metadata().edmx });
                 if (!include) queryAst = deepmerge(queryAst, this.resourcePath.ast.value.query || {});
-                await new ResourcePathVisitor(this.serverType, this.entitySets).Visit(queryAst, {}, (result || this.ctrl.prototype).elementType);
+                const lastNavigationPath = this.resourcePath.navigation[this.resourcePath.navigation.length - 1];
+                const queryType = lastNavigationPath.type == "QualifiedEntityTypeName" ?
+                    this.resourcePath.navigation[this.resourcePath.navigation.length - 1].node[ODATA_TYPE] :
+                    (result || this.ctrl.prototype).elementType;
+                await new ResourcePathVisitor(this.serverType, this.entitySets).Visit(<Token>queryAst, {}, queryType);
             }
             params[queryParam] = this.serverType.connector ? this.serverType.connector.createQuery(queryAst, elementType) : queryAst;
 
@@ -1596,7 +1600,11 @@ export class ODataProcessor extends Transform {
                 filterAst = qs.parse(filterAst).$filter;
                 if (typeof filterAst == "string") {
                     filterAst = this.serverType.parser.filter(filterAst, { metadata: this.resourcePath.ast.metadata || this.serverType.$metadata().edmx });
-                    await new ResourcePathVisitor(this.serverType, this.entitySets).Visit(<Token>filterAst, {}, (result || this.ctrl.prototype).elementType);
+                    const lastNavigationPath = this.resourcePath.navigation[this.resourcePath.navigation.length - 1];
+                    const queryType = lastNavigationPath.type == "QualifiedEntityTypeName" ?
+                        this.resourcePath.navigation[this.resourcePath.navigation.length - 1].node[ODATA_TYPE] :
+                        (result || this.ctrl.prototype).elementType;
+                    await new ResourcePathVisitor(this.serverType, this.entitySets).Visit(<Token>filterAst, {}, queryType);
                 }
             } else {
                 let token = <Token>queryString;
